refactor codebase to include example for molecules in inversion_qm9.py, general backfiltering in inversion.py, and peak finding in peak_finding.py
